bigquery: Split the testcase into unit and integration tests.
The unit tests can now be run without local GCP credentials.
